Renaming Router and Adding Destruction to list
