+++ api/.env.sample
@@ -9,3 +9,5 @@ CLIENT_ID=****
 CLIENT_SECRET=****
 ISSUER_BASE_URL=https://yukon.vivvocloud.com/
 SECRET=****
+
+V2_API_KEY_REMOTE=****
