+++ api/src/index.ts
@@ -62,7 +62,7 @@ app.use("/api/asset-owner", assetOwnerRouter);
 app.use("/api/asset-transfer", transferRouter);
 app.use("/api/mailcode", mailcodeRouter);
 app.use("/api/scan", scanRouter);
-app.use("/api/v2/asset-item", v2Router.assetTagRouterForShaun);
+app.use("/api/v2/asset-item", v2Router.assetTagRouterRemote);
 
 // serves the static files generated by the front-end
 app.use(express.static(path.join(__dirname, "web")));
