+++ api/src/routes/auth.ts
@@ -2,7 +2,7 @@ import { Express, NextFunction, Request, Response } from "express";
 import * as ExpressSession from "express-session";
 
 import { AuthUser } from "../data";
-import { AUTH_REDIRECT, FRONTEND_URL, V2_API_KEY_FOR_SHAUN } from "../config";
+import { AUTH_REDIRECT, FRONTEND_URL, V2_API_KEY_REMOTE } from "../config";
 import { UserService } from "../services";
 
 import { auth } from "express-openid-connect";
@@ -29,7 +29,7 @@ function oidcAuthenticationForEverywhereExcept(exclusions: RegExp[]) {
 }
 
 function bearerAuthentication(req: Request, res: Response, next: NextFunction) {
-    if (req.header("Authorization") === `Bearer ${V2_API_KEY_FOR_SHAUN}`) {
+    if (req.header("Authorization") === `Bearer ${V2_API_KEY_REMOTE}`) {
         return next();
     }
